feat(View): link all views & features
